requirements.txt
cycler==0.10.0
Django==2.2.17
kiwisolver==1.1.0
lxml==4.6.2
matplotlib==3.0.3
numpy==1.18.5
Pillow==7.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-docx==0.8.10
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
wordcloud==1.5.0
